Add files via upload
Certainly! Here's a short note on how to write a Python program to find happy numbers between 1 to 100:

Finding Happy Numbers in Python

Happy numbers are numbers that eventually reach 1 when replaced by the sum of the square of each digit in the number. We can create a Python program to find happy numbers between 1 and 100 using a simple algorithm.